crud

+++ src/main/java/model/PlannedTransaction.java
@@ -5,13 +5,13 @@ import java.time.LocalDate;
 public class PlannedTransaction{
     private int id;
     private String action;
-    private int money_Amount;
+    private long money_Amount;
     private User user_id;
 
     public PlannedTransaction() {
     }
 
-    public PlannedTransaction(int id, String action, int money_Amount, User user_id) {
+    public PlannedTransaction(int id, String action, long money_Amount, User user_id) {
         this.id = id;
         this.action = action;
         this.money_Amount = money_Amount;
@@ -34,11 +34,11 @@ public class PlannedTransaction{
         this.action = action;
     }
 
-    public int getMoney_Amount() {
+    public long getMoney_Amount() {
         return money_Amount;
     }
 
-    public void setMoney_Amount(int money_Amount) {
+    public void setMoney_Amount(long money_Amount) {
         this.money_Amount = money_Amount;
     }
